run postinstall in subprocess
